Move `clap` parsers into relevant modules
It's weird to have these in a separate part of the tree than where the
types they parse are defined.
