+++ src/ghci/ghci_command.rs
@@ -2,6 +2,10 @@ use std::fmt::Debug;
 use std::fmt::Display;
 use std::ops::Deref;
 
+use clap::builder::StringValueParser;
+use clap::builder::TypedValueParser;
+use clap::builder::ValueParserFactory;
+
 /// A `ghci` command.
 ///
 /// This is a string that can be written to a `ghci` session, typically a Haskell expression or
@@ -51,5 +55,32 @@ impl Deref for GhciCommand {
 
     fn deref(&self) -> &Self::Target {
         &self.0
+    }
+}
+
+/// [`clap`] parser for [`GhciCommand`] values.
+#[derive(Default, Clone)]
+pub struct GhciCommandParser {
+    inner: StringValueParser,
+}
+
+impl TypedValueParser for GhciCommandParser {
+    type Value = GhciCommand;
+
+    fn parse_ref(
+        &self,
+        cmd: &clap::Command,
+        arg: Option<&clap::Arg>,
+        value: &std::ffi::OsStr,
+    ) -> Result<Self::Value, clap::Error> {
+        self.inner.parse_ref(cmd, arg, value).map(GhciCommand)
+    }
+}
+
+impl ValueParserFactory for GhciCommand {
+    type Parser = GhciCommandParser;
+
+    fn value_parser() -> Self::Parser {
+        Self::Parser::default()
     }
 }
